requirements.txt
xgboost>=1.7
herbie-data>=2024.8
boto3>=1.28
pandas>=2.0
numpy>=1.24
rasterio>=1.3
pyproj>=3.5
scipy>=1.11
requests>=2.31
scikit-learn>=1.3
pytest>=7.4
